Be more careful about accessing attributes of the parent: if Tk has not been initialized, this will be None, but the functions will still work (there will simply be a bogus parent on the screen).  Allowing the parent to be None is useful when testing the functions from an interactive interpreter.
Add an optional keyword paramter "show" to the _QueryString class; when given
it is used to set the -show option to the entry widget.  This allows passing
show="*" or the like to askstring(), making it useful for requesting
passwords/passphrases from the user.
This closes SF bug #438517.

Changed a docstring to be less font-lock-hostile.

--HG--
branch : legacy-trunk