Apply diff2.txt from SF patch http://www.python.org/sf/565471
This patch replaces string module functions with string
methods in the Tools/world/world scripts.

It also updates two outdated URLs and the countrycodes
dictionary.

It fixes a bug where result of string.find() was checked
for truth instead of compared with -1.

It also replaces <> with != in two spots.

--HG--
branch : legacy-trunk